versionize: rework all unit test
Using a separate test crate, the current versionize-derive will use the
crate version to determine the struct version, So separate the unit tests
that depend on the crate version to the test-data crate.
