+++ crates/dbs-versionize/test-data/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "dbs-versionize-tests"
+version = "0.1.9"
+license = "Apache-2.0"
+edition = "2018"
+
+[dependencies]
+dbs-versionize = { path = "../" }
+vmm-sys-util = "0.11.0"
